tetragon: Move rawSyscallName to raw_syscall_name_linux.go
splitting syscallmetrics package into platform independent and linux specific listings
