+++ pkg/metrics/syscallmetrics/syscallmetrics.go
@@ -8,7 +8,6 @@ import (
 	"github.com/cilium/tetragon/pkg/metrics"
 	"github.com/cilium/tetragon/pkg/metrics/consts"
 	"github.com/cilium/tetragon/pkg/option"
-	"github.com/cilium/tetragon/pkg/syscallinfo"
 	"github.com/prometheus/client_golang/prometheus"
 )
 
@@ -62,24 +61,4 @@ func Handle(event interface{}) {
 		processLabels := option.CreateProcessLabels(namespace, workload, pod, binary)
 		syscallStats.WithLabelValues(processLabels, syscall).Inc()
 	}
-}
-
-func rawSyscallName(tp *tetragon.ProcessTracepoint) string {
-	sysID := int64(-1)
-	var sysABI string
-	for _, x := range tp.Args {
-		if x, ok := x.GetArg().(*tetragon.KprobeArgument_SyscallId); ok {
-			sysID = int64(x.SyscallId.Id)
-			sysABI = x.SyscallId.Abi
-			break
-		}
-	}
-	if sysID == -1 {
-		return ""
-	}
-	name, err := syscallinfo.GetSyscallName(sysABI, int(sysID))
-	if err != nil {
-		return ""
-	}
-	return name
 }
